Update branch with UUID n1nsy1feajq6x0dfr2zn1bi6l
New branch nickname: mdarc4

Rename project from "mdfast" into "mdarc4"

The reason is that a speed comparison unfortunately showed that md5sum
was much faster than this simplistic algorithm.

A name containing "fast" was therefore no longer appropriate.

However, the advantage of a trivally simple implementation still
applies.

This algorithm can easily be re-implemented from memory, which is
certainly not possible for MD5 or any other known hash algorithm which
is able to generate a 256-bit digest.

It is also simpler than a simple CRC, because no polynomial and
implementation parameters (reversed, reflected etc) needs to be
remembered.

What needs to be remembered is:

* The standard ARCFOUR algorithm
* That all key octets are processed exactly once
* The first 3072 output octets are discarded
* The next 256 bits are used as the hash value
* When converting to base-32, use alphabet A-Z 0-9 without 0, 1, O, I.