Support command line arguments
All non-option arguments are expected to be pathnames of binary files to
be hashed. In this case, the output for each file will consist of its
message digest in base-32 format and the path name, separated by a
single space.

Only a single option is supported: "--" which stops processing of
further options.

Options are only recognized before all non-option arguments.

Without any options, the program behaves as before and just hashes its
standard input stream.