Update branch with UUID n1nsy1feajq6x0dfr2zn1bi6l
New branch nickname: mdrc4

Rename project

The previous name "mdarc4" was used because RC4 is a registered
tradename where ARCFOUR isn't.

However, "mdrc4" is not the same as "RC4", and therefore the tradename
issue should not apply.

The algorithm itself is always referred to as ARCFOUR in the
documentation and source code, never as "RC4".