Update branch with UUID n1nsy1feajq6x0dfr2zn1bi6l
New branch nickname: rc4hash

Rename project

As my suspicions about the cryptographic security of the algorithm have
deepened, I decided it was no longer justified to call it a "message
digest" algorithm, because such algorithms are usually cryptographically
secure.

It remains to be an interesting and easy-to-implement hash algorithm
with arbitrarily long output though, and the new name accounts for this.

It might also be acceptable to shorten the -drop3072 into -drop256 if no
cryptographic security was provided anyway, therefore considerably
speeding up the finalization phase of the hash calculation.

However, contrary to my doubts I have no hard proof yet that the
algorithm is indeed cryptographically insecure, and reducing the
post-processing effort could considerably undermine its security then.
Therefore leaving the post-processing as is for now.