Revert key setup modifications and SXS-mode
Although not necessary for security purposes, revert the simplifications
to the ARCFOUR key setup and use a standard setup instead. This allows
to verify the implementation by comparing standard test vectors.

Drop the SXS-mode and use normal XOR stream-cipher processing. The SXS
mode only made a difference if the same key was used for encrypting more
than a single message, which must be avoided anyway.

Rename the utility into rc4drop3072, because it is actually implements
that algorithm now without any deviations ("improvements").

Drop the support for command-line options. Instead, feed everything
needed via standard input. Define a format for that and document it in
the usage help of the respective utility.

Add initial implementation of the treyfer block cipher in OFB mode,
turning it into a stream cipher.

The idea is to combine rc4drop3072 and treyfer-ofb by chaining, making a
more secure cipher than ARCFOUR alone.