Make ARCFOUR encryption utility more versatile
Instead of hard-coding a particular "drop-N" parameter, accept the
parameter value from the input specification.

This also allows the algorithm to be degraded to ARCFOUR-drop0 which is
identical to the original ARCFOUR algorithm.

Also support different key sizes in the same way as the original
algorithm, providing the actual key size as another parameter value from
the input specification.

As a result of those changes, key sizes from 1 through 256 octets are
now supported.

The only difference is that longer keys than 256 octets are not
supported in this implementation, while the original algorithm supports
keys of any size. However, it "supports" keys longer than 256 octets
only be ignoring the remaining octets of the key.

Which means truncating longer keys to 256 bytes will give identical
results for the original algorithm and the implementation provided in
this project.