Add online-help to rc4hash
The -h and -V options shall display help and version information.

Include an option for directly specifying the desired output width of
the hash digest as a number of characters.

This is preferable to the existing specification as a desired number of
bits or bytes in cases where the output width of the hash has to be
chosen in order to fit into a predefined column layout for publishing.