Fix severe implementation errors
When trying to verify the implementation using published ARCFOUR test
vectors, I detected that the key setup is completely wrong.

First of all, the loop was run too few times because the termination
criterion was wrong. The "-drop"-parameter was also interpreted
incorrectly.

Next, the recycling did not work correctly for the first byte of the
repeated key.

And thirdly, the setup algorithm works differently than I assumed!

The latter problem will be fixed later - fix the former ones instead.

Also add -h and -V options to 'sxs-crypt' utility.