New utilities for bitwise XOR and decimal undump
The new xor10 can be used to combine the output of rc4csprng into the
actual result of the ARCFOUR stream cipher, although still in decimal
format.

Finally the new dec2raw can be used to convert the decimal format back
to raw binary bytes, representing the encrypted/decrypted output.