Fix key setup plagued by incorrect description
The correct rules are: In key setup, index1 is incremented at the end of
the loop. In PRNG generation mode, index1 is incremented at the
beginning of the loop.

I don't think that changing the order in the previous versions had any
effect on the security of the algorithm, but it definitely gave
different results.