Define feature test macro for 64 bit file offsets
It turned out that without the symbol _FILE_OFFSET_BITS defined as 64,
the fopen() function cannot open files larger than 4 GiB!

I would understand if this happened when I was trying to use fseek() on
such a file. But I did not. Just plain old fopen() and sequential
reading. Nevertheless it did not work.

This is really annoying.

Also, _FILE_OFFSET_BITS seems to be glibc or Linux specific. The POSIX
standard does not definie sich a feature test macro.