Display usage also if required options are missing
Also change -e and -d into -E and -D in order to indicate this is not
high-level encryption/decryption but rather "raw" encryption/decryption.

High-level encryption should automatically prepend a salt and append a
MAC. High-level decryption should automatically extract the salt, derive
the raw decryption keys from it, and verify the MAC also.

The current rc4sxs-crypt does not do that, and hence the emphasis on
"raw".