Change usage documentation and add s-box backend
The s-box backend can be used instead of the xor-backend for actually
carrying out encryption or decryption using the already-provided
pseudorandom stream.

The idea of the s-box backend is that a reversible substitution is much
better than an XOR operation, because bit-flipping attacks will no
longer be possible.

Therefore, an s-box is initialized by a the first part of the output of
the csprng, and then every plaintext value is transformed by
substituting it. Every s-box table slot is pseudo-randomly swapped
before proceding, using the next pseudorandom value produced by the
csprng.

Reverting this process require setting up and updating a reverse lookup
table, making it a more expensive operation as the basic transformation.

Therefore, the basic transformation is used for decryption, and the more
expensive one for encryption.