Change key size to match block size
I finally was able to get hold of the initial publishing paper of the
Treyfer cipher.

Studying this paper made it clear that the factor of 8 used for the
rounds in the Wikipedia version is not because of rotating every bit of
each octet.

Instead it is because of the key and block size, which is 8 bytes in the
original version: Every data octet and key octet shall be processed once
per round.

This also means that the key size must be the same as the block size
when scaling this scheme.

Therefore, the local implementation has been changed to use both a 512
bit block and key.

If necessary, a given 256 bit key can always be stretched/padded to 512
bits somehow.

Also redesign the s-box array initializers.

Providing the calculated octet values as octal-encoded character
literals did not prevent gcc from showing warning messages with maximum
warnings enabled.

Use macros for casting the integer values into character constants
cleanly.