Add amalgamation template
This script can unpack a couple of other scripts which need one another
for running into a temporary directory and run the first of those
script, passing it any command line arguments passed to the amalgamation
wrapper script.

Once that first script returns, remove the temporary directory and of
the scripts within it and return the result code of the first script.

The other scripts are never called by the amalgamation wrapper, but it
is assumed that the first script calls them, using the same directory
path as itself for the invocations.